requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
apipkg==1.5               # via execnet
attrs==20.3.0             # via pytest
babel==2.9.0              # via flask-babel
bcrypt==3.2.0             # via flask-bcrypt
cffi==1.14.3              # via bcrypt
click==7.1.2              # via flask
coverage==5.3             # via -r requirements.in
dnspython==2.0.0          # via email-validator
email-validator==1.1.2    # via wtforms
execnet==1.7.1            # via pytest-xdist
flask-admin==1.5.7        # via -r requirements.in
flask-babel==2.0.0        # via -r requirements.in
flask-bcrypt==0.7.1       # via -r requirements.in
flask-login==0.5.0        # via -r requirements.in
flask-reuploaded==0.3.2   # via -r requirements.in
flask-sqlalchemy==2.4.4   # via -r requirements.in
flask-wtf==0.14.3         # via -r requirements.in
flask==1.1.2              # via -r requirements.in, flask-admin, flask-babel, flask-bcrypt, flask-login, flask-reuploaded, flask-sqlalchemy, flask-wtf
gunicorn==20.0.4          # via -r requirements.in
idna==2.10                # via email-validator
iniconfig==1.1.1          # via pytest
itsdangerous==1.1.0       # via flask, flask-wtf
jinja2==2.11.2            # via flask, flask-babel
markdown==3.3.3           # via -r requirements.in
markupsafe==1.1.1         # via jinja2, wtforms
packaging==20.4           # via pytest
phonenumbers==8.12.13     # via -r requirements.in
pillow==8.0.1             # via -r requirements.in
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest, pytest-forked
pycparser==2.20           # via cffi
pyparsing==2.4.7          # via packaging
pytest-forked==1.3.0      # via pytest-xdist
pytest-xdist==2.1.0       # via -r requirements.in
pytest==6.1.2             # via -r requirements.in, pytest-forked, pytest-xdist
python-dateutil==2.8.1    # via vobject
pytz==2020.4              # via babel, flask-babel
six==1.15.0               # via bcrypt, packaging, python-dateutil
sqlalchemy==1.3.20        # via flask-sqlalchemy
toml==0.10.2              # via pytest
vobject==0.9.6.1          # via -r requirements.in
werkzeug==1.0.1           # via flask
wtforms[email]==2.3.3     # via -r requirements.in, flask-admin, flask-wtf

# The following packages are considered to be unsafe in a requirements file:
# setuptools
